Final Game with 2 maps

+++ Eksamensprojekt/inc/Collider.h
@@ -38,7 +38,7 @@ struct player{
     int8_t lives;
     uint8_t direction;
     uint8_t bulletType;
-    int64_t score;
+    int32_t score;
     struct bullet shots[3];
 };
